Consolidate listing steps
