+++ src/go/cukestep/local_repo_steps.go
@@ -2,14 +2,16 @@ package cukestep
 
 import (
 	"context"
-	"fmt"
 
 	"github.com/cucumber/godog"
 	support "github.com/kkrull/marmot/cukesupport"
 )
 
-// State to clear between scenarios
+/* State */
+
 var thoseLocalRepositories *support.LocalRepositories = support.NoLocalRepositories()
+
+/* Configuration */
 
 // Add step definitions related to repositories on the local filesystem.
 func AddLocalRepositorySteps(ctx *godog.ScenarioContext) {
@@ -28,6 +30,8 @@ func AddLocalRepositorySteps(ctx *godog.ScenarioContext) {
 	ctx.Then(`^that repository listing should include those local repositories$`, thatListingShouldHaveLocals)
 }
 
+/* Steps */
+
 func createLocalGitRepositories() error {
 	if repoDir, pathErr := support.TestSubDir("empty-dir"); pathErr != nil {
 		return pathErr
@@ -35,19 +39,6 @@ func createLocalGitRepositories() error {
 		return repoErr
 	} else {
 		thoseLocalRepositories = support.SomeLocalRepositories(repo)
-		return nil
-	}
-}
-
-func listLocal() error {
-	if factory, configErr := support.ThatQueryFactory(); configErr != nil {
-		return fmt.Errorf("repository_steps: failed to configure; %w", configErr)
-	} else if listRepositories, appErr := factory.NewListLocalRepositories(); appErr != nil {
-		return fmt.Errorf("repository_steps: failed to initialize; %w", appErr)
-	} else if repositories, runErr := listRepositories(); runErr != nil {
-		return fmt.Errorf("repository_steps: failed to run query; %w", runErr)
-	} else {
-		thatListing = repositories
 		return nil
 	}
 }
